multimodal-generation/repeatable-patterns/03-chat-based-image-gen/python/requirements.txt
appnope==0.1.4
asttokens==3.0.0
blessed==1.20.0
boto3==1.37.28
botocore==1.37.28
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
editor==1.6.6
executing==2.2.0
inquirer==3.4.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pyzmq==26.4.0
readchar==4.2.1
runs==1.2.2
s3transfer==0.11.4
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
urllib3==2.3.0
wcwidth==0.2.13
xmod==1.8.1
